Moved auth token class
