+++ pnnl.goss.core/src/pnnl/goss/core/security/impl/SecurityConfigImpl.java
@@ -24,6 +24,7 @@ import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import com.northconcepts.exception.SystemException;
 
+import pnnl.goss.core.security.JWTAuthenticationToken;
 import pnnl.goss.core.security.SecurityConfig;
 import pnnl.goss.core.security.SecurityConstants;
